+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.4
 
 * [CHANGE] Update system container images to use GitHub registry to avoid rate limits with Docker Hub.
+* [CHANGE] Update borg to use `--numeric-ids` instead of the deprecated `--numeric-owner`.
 
 ***